Change OutIn from byte array to double.

+++ IOModule.cs
@@ -5,7 +5,7 @@ namespace BrainPad {
         public virtual double In() => 0;
         public virtual void Out(double data) { }
         public virtual void Out(double[] data) { }
-        public virtual double OutIn(byte[] data, byte[] result) => 0;
+        public virtual double OutIn(double[] data, double[] result) => 0;
         public abstract void Dispose();
     }
 }
